import-url: reversing cosmetic changes
